Implemented handling if pid_table contains record in the beginning of ./bp

+++ bp/configstorage.cpp
@@ -80,7 +80,7 @@ unsigned int ConfigStorage::getRerunAfter() const
     return rerunAfter;
 }
 
-int ConfigStorage::getIdOfPid() const
+long ConfigStorage::getIdOfPid() const
 {
     return idOfPid;
 }
