+++ bp/configstorage.h
@@ -25,7 +25,7 @@ private:
     ILogger* logger = nullptr;
     int rerunTimes;
     unsigned int rerunAfter;
-    int idOfPid;
+    long idOfPid;
 public:
     ConfigStorage(ConfigParser* parser);
     ~ConfigStorage();
@@ -44,7 +44,7 @@ public:
     int getPooledConnections() const;
     int getRerunTimes() const;
     unsigned int getRerunAfter() const;
-    int getIdOfPid() const;
+    long getIdOfPid() const;
 };
 
 #endif // CONFIGSTORAGE_H
